requirements.txt
Django>=2.2
django-filter>=2.1.0
djangorestframework>=3.9.2
